+++ providers/gcp/CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## [0.0.2](https://github.com/open-feature/java-sdk-contrib/compare/dev.openfeature.contrib.providers.gcp-v0.0.1...dev.openfeature.contrib.providers.gcp-v0.0.2) (2026-06-15)
+
+
+### ✨ New Features
+
+* add GCP Secret Manager OpenFeature provider ([#1772](https://github.com/open-feature/java-sdk-contrib/issues/1772)) ([17de2b7](https://github.com/open-feature/java-sdk-contrib/commit/17de2b76ecb864e93dc42b53e1d72dd08f4a5bc3))
+* **gcp:** add parameter manager support ([#1808](https://github.com/open-feature/java-sdk-contrib/issues/1808)) ([f3df150](https://github.com/open-feature/java-sdk-contrib/commit/f3df150387036572740d2fd3fa22d3069a1b730e))
+
 ## 0.0.1
 
 ### ✨ New Features